+++ src/Integrations/WebhookManager/WebhookManagerBridge.php
@@ -53,7 +53,7 @@ class WebhookManagerBridge
 
     public static function available(): bool
     {
-        return (bool) config('statamic-offers.integrations.webhook_manager', true)
+        return (bool) config('statamic-offers.webhook_manager.enabled', true)
             && class_exists(self::FACADE)
             && interface_exists(self::TRIGGER_INTERFACE);
     }